+++ plugins/woocommerce-admin/packages/components/src/chart/d3chart/utils/axis.js
@@ -197,7 +197,7 @@ export const drawAxis = ( node, params, xOffset ) => {
 			: compareStrings( formatter( prevMonth ), formatter( monthDate ) ).join( ' ' );
 	};
 
-	const yGrids = getYGrids( params.yMax );
+	const yGrids = getYGrids( params.yMax === 0 ? 1 : params.yMax );
 
 	const ticks = params.xTicks.map( d => ( params.type === 'line' ? moment( d ).toDate() : d ) );
 
@@ -257,7 +257,7 @@ export const drawAxis = ( node, params, xOffset ) => {
 		.attr( 'text-anchor', 'start' )
 		.call(
 			d3AxisLeft( params.yTickOffset )
-				.tickValues( yGrids )
+				.tickValues( params.yMax === 0 ? [ yGrids[ 0 ] ] : yGrids )
 				.tickFormat( d => params.yFormat( d !== 0 ? d : 0 ) )
 		);