+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.5.0 (unreleased)
+- Improves display of charts where all values are 0.
 - Fix X-axis labels in hourly bar charts.
 - New `<Search>` prop named `showClearButton`, that will display a 'Clear' button when the search box contains one or more tags.